MQE-1374: waitForPageLoads being inserted unnecessarily
- Add unit test
